Change explosion in UI and cgange Readme

+++ src/bomb.js
@@ -1,9 +1,9 @@
 import * as PIXI from 'pixi.js';
-import { app, bombState} from './index.js';
-import { sizeRect } from './field.js';
+import { app, bombState, gameState} from './index.js';
+import { fieldSize, sizeRect, widthField } from './field.js';
 import { bombermen, handlePlayerDestroy } from './bomberman.js';
 import { arrStone } from './stone.js';
-import { handleWallDestroy } from './wall.js';
+import { arrWall, handleWallDestroy } from './wall.js';
 import { checkedEnemy, handleEnemyDestroy } from './enemy.js';
 import { handleBonusDestroy, heandleTeleportExplosion } from './bonuses.js';
 
@@ -62,13 +62,56 @@ const handleBombExplosion = (bomb) => {
     
 }   
 
+// const createExplosionX = (bomb) => {
+//     const explosionContainer = new PIXI.Container();
+//     const explosion = new PIXI.Sprite(explosionSprite);
+//     explosion.width = sizeRect * bombState.explosionSize;
+//     explosion.height = sizeRect;
+//     explosionContainer.zIndex = 2;
+//     explosion.position.x = bomb.position.x - ((sizeRect - bomb.width) / 2) - sizeRect * bombState.bombRadius;
+//     explosion.position.y = bomb.position.y - ((sizeRect - bomb.height));
+    
+//     setTimeout(() => {
+//         explosion.destroy({children: true})
+//     }, 500)
+
+//     explosionContainer.addChild(explosion);
+//     app.stage.addChild(explosionContainer);
+// }
+
 const createExplosionX = (bomb) => {
     const explosionContainer = new PIXI.Container();
     const explosion = new PIXI.Sprite(explosionSprite);
-    explosion.width = sizeRect * bombState.explosionSize;
+    let widthX = 1;
+    let biasX = 0;
+    for(let i = 1; i <= bombState.bombRadius; i++) {
+        if(arrStone[bomb.bombIndex + i] || Math.floor((bomb.bombIndex + i) / widthField) !== Math.floor(bomb.bombIndex / widthField)){
+            break;
+        }
+        if(arrWall[bomb.bombIndex + i]) {
+            widthX += 1;
+            break;
+        }
+        widthX += 1;
+    }
+
+    for(let i = 1; i <= bombState.bombRadius; i++) {
+        if(arrStone[bomb.bombIndex - i] || Math.floor((bomb.bombIndex - i) / widthField) !== Math.floor(bomb.bombIndex / widthField)){
+            break;
+        }
+        if(arrWall[bomb.bombIndex - i]) {
+            widthX += 1;
+            biasX += 1;
+            break;
+        }
+        widthX += 1;
+        biasX += 1;
+    }
+
+    explosion.width = sizeRect * widthX;
     explosion.height = sizeRect;
     explosionContainer.zIndex = 2;
-    explosion.position.x = bomb.position.x - ((sizeRect - bomb.width) / 2) - sizeRect * bombState.bombRadius;
+    explosion.position.x = bomb.position.x - (sizeRect - bomb.width) / 2 - (sizeRect * biasX);
     explosion.position.y = bomb.position.y - ((sizeRect - bomb.height));
     
     setTimeout(() => {
@@ -79,14 +122,61 @@ const createExplosionX = (bomb) => {
     app.stage.addChild(explosionContainer);
 }
 
+// const createExplosionY = (bomb) => {
+//     const explosionContainer = new PIXI.Container();
+//     const explosion = new PIXI.Sprite(explosionSprite);
+//     explosion.width = sizeRect;
+//     explosion.height = sizeRect * bombState.explosionSize;
+//     explosionContainer.zIndex = 2;
+//     explosion.position.x = bomb.position.x - ((sizeRect - bomb.width) / 2);
+//     explosion.position.y = bomb.position.y - ((sizeRect - bomb.height)) - sizeRect * bombState.bombRadius;
+    
+//     setTimeout(() => {
+//         explosion.destroy({children: true})
+//     }, 500)
+
+//     explosionContainer.addChild(explosion);
+//     app.stage.addChild(explosionContainer);
+// }
+
 const createExplosionY = (bomb) => {
     const explosionContainer = new PIXI.Container();
     const explosion = new PIXI.Sprite(explosionSprite);
+    let widthY = 1;
+    let biasY = 0;
+    for(let i = 1; i <= bombState.bombRadius; i++) {
+        if(arrStone[bomb.bombIndex - widthField * i] || bomb.bombIndex - widthField * i < 0){
+            break;
+        }
+        if(arrWall[bomb.bombIndex - widthField * i]) {
+            widthY += 1;
+            biasY += 1;
+            break;
+        }
+        widthY += 1;
+        biasY += 1;
+    }
+
+    for(let i = 1; i <= bombState.bombRadius; i++) {
+        if(arrStone[bomb.bombIndex + widthField * i] || bomb.bombIndex + widthField * i >= fieldSize){
+            break;
+        }
+        if(arrWall[bomb.bombIndex + widthField * i]) {
+            widthY += 1;
+            break;
+        }
+        widthY += 1;
+    }
+
+    if(widthY === 1) {
+        return;
+    }
+
     explosion.width = sizeRect;
-    explosion.height = sizeRect * bombState.explosionSize;
+    explosion.height = sizeRect * widthY;
     explosionContainer.zIndex = 2;
-    explosion.position.x = bomb.position.x - ((sizeRect - bomb.width) / 2);
-    explosion.position.y = bomb.position.y - ((sizeRect - bomb.height)) - sizeRect * bombState.bombRadius;
+    explosion.position.x = bomb.position.x - (sizeRect - bomb.width) / 2;
+    explosion.position.y = bomb.position.y - ((sizeRect - bomb.height)) - (sizeRect * biasY);
     
     setTimeout(() => {
         explosion.destroy({children: true})
